Refactoring training modules
