+++ nmt_chainer/__main__.py
@@ -1,12 +1,12 @@
 #!/usr/bin/env python -O
 
 import argparse
-import nmt_chainer.training_module.train as train
+import nmt_chainer.training_module.train_config as train
 import nmt_chainer.eval as eval_module
 import nmt_chainer.make_data as make_data
 import nmt_chainer.server as server
 
-import nmt_chainer.utilities.utils as utils_command
+import nmt_chainer.utils_command as utils_command
 
 from nmt_chainer.utilities import versioning_tools
 import sys
