+++ tests/macro_test.py
@@ -6,10 +6,13 @@ __version__ = "1.0"
 __email__ = "fabien.cromieres@gmail.com"
 __status__ = "Development"
 
-import nmt_chainer.make_data as make_data
-import nmt_chainer.training_module.train as train
-import nmt_chainer.eval as eval
-import nmt_chainer.utilities.utils as utils
+# import nmt_chainer.make_data as make_data
+# import nmt_chainer.training_module.train as train
+# import nmt_chainer.eval as eval
+# import nmt_chainer.utilities.utils as utils
+
+from nmt_chainer.__main__ import main
+
 import os.path
 import pytest
 
@@ -26,21 +29,21 @@ class TestMacro:
         train_prefix = str(train_dir.join("test1.train"))
         data_src_file = os.path.join(test_data_dir, "src2.txt")
         data_tgt_file = os.path.join(test_data_dir, "tgt2.txt")
-        args = '{0} {1} {2} --dev_src {0} --dev_tgt {1}'.format(
+        args = 'make_data {0} {1} {2} --dev_src {0} --dev_tgt {1}'.format(
             data_src_file, data_tgt_file, data_prefix).split(' ')
-        make_data.cmdline(arguments = args)
+        main(arguments = args)
         
-        args_train = [data_prefix, train_prefix] + "--max_nb_iters 1000 --mb_size 2 --Ei 10 --Eo 12 --Hi 30 --Ha 70 --Ho 15 --Hl 23".split(" ")
+        args_train = ["train", data_prefix, train_prefix] + "--max_nb_iters 1000 --mb_size 2 --Ei 10 --Eo 12 --Hi 30 --Ha 70 --Ho 15 --Hl 23".split(" ")
         if gpu is not None:
             args_train += ['--gpu', gpu]
-        train.command_line(arguments = args_train)
+        main(arguments = args_train)
 
         eval_dir = tmpdir.mkdir("eval")
         translation_file = os.path.join(str(eval_dir), 'translations.txt')
-        args_eval = [train_prefix + '.train.config', train_prefix + '.model.best.npz', data_src_file, translation_file] + '--mode beam_search --beam_width 30'.split(' ') 
+        args_eval = ["eval", train_prefix + '.train.config', train_prefix + '.model.best.npz', data_src_file, translation_file] + '--mode beam_search --beam_width 30'.split(' ') 
         if gpu is not None:
             args_eval += ['--gpu', gpu]
-        eval.command_line(arguments = args_eval)
+        main(arguments = args_eval)
 
         with open(data_tgt_file) as f:
             expected_translations = f.readlines()
@@ -69,21 +72,21 @@ class TestMacro:
             train_prefix = str(train_dir.join("test1.train"))
             data_src_file = os.path.join(test_data_dir, "src2.txt")
             data_tgt_file = os.path.join(test_data_dir, "tgt2.txt")
-            args = '{0} {1} {2} --dev_src {0} --dev_tgt {1}'.format(
+            args = 'make_data {0} {1} {2} --dev_src {0} --dev_tgt {1}'.format(
                 data_src_file, data_tgt_file, data_prefix).split(' ')
-            make_data.cmdline(arguments = args)
+            main(arguments = args)
             
-            args_train = [data_prefix, train_prefix] + "--max_nb_iters 200 --mb_size 2 --Ei 10 --Eo 12 --Hi 30 --Ha 70 --Ho 15 --Hl 23".split(" ")
+            args_train = ["train", data_prefix, train_prefix] + "--max_nb_iters 200 --mb_size 2 --Ei 10 --Eo 12 --Hi 30 --Ha 70 --Ho 15 --Hl 23".split(" ")
             if gpu is not None:
                 args_train += ['--gpu', gpu]
-            train.command_line(arguments = args_train)
+            main(arguments = args_train)
 
             beam_search_eval_dir = tmpdir.mkdir("eval_beam_search_{0}".format(i))
             beam_search_file = os.path.join(str(beam_search_eval_dir), 'translations.txt')
-            args_eval = [train_prefix + '.train.config', train_prefix + '.model.best.npz', data_src_file, beam_search_file] + '--mode beam_search --beam_width 1'.split(' ') 
+            args_eval = ["eval", train_prefix + '.train.config', train_prefix + '.model.best.npz', data_src_file, beam_search_file] + '--mode beam_search --beam_width 1'.split(' ') 
             if gpu is not None:
                 args_eval += ['--gpu', gpu]
-            eval.command_line(arguments = args_eval)
+            main(arguments = args_eval)
 
             greedy_search_eval_dir = tmpdir.mkdir("eval_greedy_search_{0}".format(i))
             greedy_search_file = os.path.join(str(greedy_search_eval_dir), 'translations.txt')
@@ -116,29 +119,29 @@ class TestMacro:
         train_prefix = str(train_dir.join("test1.train"))
         data_src_file = os.path.join(test_data_dir, "src2.txt")
         data_tgt_file = os.path.join(test_data_dir, "tgt2.txt")
-        args = '{0} {1} {2} --dev_src {0} --dev_tgt {1}'.format(
+        args = 'make_data {0} {1} {2} --dev_src {0} --dev_tgt {1}'.format(
             data_src_file, data_tgt_file, data_prefix).split(' ')
-        make_data.cmdline(arguments = args)
+        main(arguments = args)
         
-        args_train = [data_prefix, train_prefix] + "--max_nb_iters 200 --mb_size 2 --Ei 10 --Eo 12 --Hi 30 --Ha 70 --Ho 15 --Hl 23".split(" ")
+        args_train = ["train", data_prefix, train_prefix] + "--max_nb_iters 200 --mb_size 2 --Ei 10 --Eo 12 --Hi 30 --Ha 70 --Ho 15 --Hl 23".split(" ")
         if gpu is not None:
             args_train += ['--gpu', gpu]
-        train.command_line(arguments = args_train)
+        main(arguments = args_train)
 
         beam_search_eval_dir = tmpdir.mkdir("eval_beam_search")
         beam_search_file = os.path.join(str(beam_search_eval_dir), 'translations.txt')
-        args_eval = [train_prefix + '.train.config', train_prefix + '.model.best.npz', data_src_file, beam_search_file] + '--mode beam_search --beam_width 30'.split(' ') 
+        args_eval = ["eval", train_prefix + '.train.config', train_prefix + '.model.best.npz', data_src_file, beam_search_file] + '--mode beam_search --beam_width 30'.split(' ') 
         if gpu is not None:
             args_eval += ['--gpu', gpu]
-        eval.command_line(arguments = args_eval)
+        main(arguments = args_eval)
 
         ensemble_search_eval_dir = tmpdir.mkdir("eval_ensemble_search")
         ensemble_search_file = os.path.join(str(ensemble_search_eval_dir), 'translations.txt')
-        args_eval = [train_prefix + '.train.config', train_prefix + '.model.best.npz', data_src_file, ensemble_search_file] + \
+        args_eval = ["eval", train_prefix + '.train.config', train_prefix + '.model.best.npz', data_src_file, ensemble_search_file] + \
             '--mode beam_search --beam_width 30 --additional_training_config {0} {0} --additional_trained_model {1} {1}'.format(train_prefix + '.train.config', train_prefix + '.model.best.npz').split(' ') 
         if gpu is not None:
             args_eval += ['--gpu', gpu]
-        eval.command_line(arguments = args_eval)
+        main(arguments = args_eval)
 
         with open(beam_search_file) as f:
             beam_search_translations = f.readlines()
@@ -167,23 +170,23 @@ class TestMacro:
             train_prefix = str(train_dir.join("test1.train"))
             data_src_file = os.path.join(test_data_dir, "src2.txt")
             data_tgt_file = os.path.join(test_data_dir, "tgt2.txt")
-            args = '{0} {1} {2} --dev_src {0} --dev_tgt {1}'.format(
+            args = 'make_data {0} {1} {2} --dev_src {0} --dev_tgt {1}'.format(
                 data_src_file, data_tgt_file, data_prefix).split(' ')
-            make_data.cmdline(arguments = args)
+            main(arguments = args)
             
-            args_train = [data_prefix, train_prefix] + "--max_nb_iters 200 --mb_size 2 --Ei 10 --Eo 12 --Hi 30 --Ha 70 --Ho 15 --Hl 23".split(" ")
+            args_train = ["train", data_prefix, train_prefix] + "--max_nb_iters 200 --mb_size 2 --Ei 10 --Eo 12 --Hi 30 --Ha 70 --Ho 15 --Hl 23".split(" ")
             if gpu is not None:
                 args_train += ['--gpu', gpu]
-            train.command_line(arguments = args_train)
+            main(arguments = args_train)
 
         train_dir = str(tmpdir.join("train_0"))
         train_prefix = os.path.join(train_dir, "test1.train")
         beam_search_eval_dir = tmpdir.mkdir("eval_beam_search")
         beam_search_file = os.path.join(str(beam_search_eval_dir), 'translations.txt')
-        args_eval_beam_search = [train_prefix + '.train.config', train_prefix + '.model.best.npz', data_src_file, beam_search_file] + '--mode beam_search --beam_width 30'.split(' ') 
+        args_eval_beam_search = ["eval", train_prefix + '.train.config', train_prefix + '.model.best.npz', data_src_file, beam_search_file] + '--mode beam_search --beam_width 30'.split(' ') 
         if gpu is not None:
             args_eval_beam_search += ['--gpu', gpu]
-        eval.command_line(arguments = args_eval_beam_search)
+        main(arguments = args_eval_beam_search)
 
         ensemble_search_eval_dir = tmpdir.mkdir("eval_ensemble_search")
         ensemble_search_file = os.path.join(str(ensemble_search_eval_dir), 'translations.txt')
@@ -193,12 +196,12 @@ class TestMacro:
         train_prefix_2 = os.path.join(train_dir_2, "test1.train")
         train_dir_3 = str(tmpdir.join("train_3"))
         train_prefix_3 = os.path.join(train_dir_3, "test1.train")
-        args_eval_ensemble_search = [train_prefix_1 + '.train.config', train_prefix_1 + '.model.best.npz', data_src_file, ensemble_search_file] + \
+        args_eval_ensemble_search = ["eval", train_prefix_1 + '.train.config', train_prefix_1 + '.model.best.npz', data_src_file, ensemble_search_file] + \
             '--mode beam_search --beam_width 30 --additional_training_config {0} {1} --additional_trained_model {2} {3}'.format(
                 train_prefix_2 + '.train.config', train_prefix_3 + '.train.config', train_prefix_2 + '.model.best.npz', train_prefix_3 + '.model.best.npz').split(' ') 
         if gpu is not None:
             args_eval_ensemble_search += ['--gpu', gpu]
-        eval.command_line(arguments = args_eval_ensemble_search)
+        main(arguments = args_eval_ensemble_search)
 
         with open(beam_search_file) as f:
             beam_search_translations = f.readlines()
