+++ tests/training_management_test.py
@@ -6,9 +6,10 @@ __version__ = "1.0"
 __email__ = "fabien.cromieres@gmail.com"
 __status__ = "Development"
 
-import nmt_chainer.make_data as make_data
-import nmt_chainer.training_module.train as train
-import nmt_chainer.eval as eval
+# import nmt_chainer.make_data as make_data
+# import nmt_chainer.training_module.train as train
+# import nmt_chainer.eval as eval
+from nmt_chainer.__main__ import main
 import os.path
 import pytest
 
@@ -24,14 +25,14 @@ class TestTrainingManagement:
         train_prefix = str(train_dir.join("test1.train"))
         data_src_file = os.path.join(test_data_dir, "src2.txt")
         data_tgt_file = os.path.join(test_data_dir, "tgt2.txt")
-        args = '{0} {1} {2} --dev_src {0} --dev_tgt {1}'.format(
+        args = 'make_data {0} {1} {2} --dev_src {0} --dev_tgt {1}'.format(
             data_src_file, data_tgt_file, data_prefix).split(' ')
-        make_data.cmdline(arguments = args)
+        main(arguments = args)
         
-        args_train = [data_prefix, train_prefix] + "--max_nb_iters 10 --mb_size 2 --Ei 10 --Eo 12 --Hi 30 --Ha 70 --Ho 15 --Hl 23 --save_ckpt_every 5".split(" ")
+        args_train = ["train", data_prefix, train_prefix] + "--max_nb_iters 10 --mb_size 2 --Ei 10 --Eo 12 --Hi 30 --Ha 70 --Ho 15 --Hl 23 --save_ckpt_every 5".split(" ")
         if gpu is not None:
             args_train += ['--gpu', gpu]
-        train.command_line(arguments = args_train)
+        main(arguments = args_train)
         
     def test_config_saving(self, tmpdir, gpu):
         """
@@ -43,23 +44,23 @@ class TestTrainingManagement:
         train_prefix = str(train_dir.join("test1.train"))
         data_src_file = os.path.join(test_data_dir, "src2.txt")
         data_tgt_file = os.path.join(test_data_dir, "tgt2.txt")
-        args = '{0} {1} {2} --dev_src {0} --dev_tgt {1}'.format(
+        args = 'make_data {0} {1} {2} --dev_src {0} --dev_tgt {1}'.format(
             data_src_file, data_tgt_file, data_prefix).split(' ')
-        make_data.cmdline(arguments = args)
+        main(arguments = args)
         
-        args_train = [data_prefix, train_prefix] + "--max_nb_iters 5 --mb_size 2 --Ei 10 --Eo 12 --Hi 30 --Ha 70 --Ho 15 --Hl 23".split(" ")
+        args_train = ["train", data_prefix, train_prefix] + "--max_nb_iters 5 --mb_size 2 --Ei 10 --Eo 12 --Hi 30 --Ha 70 --Ho 15 --Hl 23".split(" ")
         if gpu is not None:
             args_train += ['--gpu', gpu]
-        train.command_line(arguments = args_train)
+        main(arguments = args_train)
         
         config_filename = train_prefix + ".train.config"
         
         train_prefix_2 = train_prefix + ".2"
-        args_train = ["--config", config_filename, "--save_prefix", train_prefix_2]
+        args_train = ["train", "--config", config_filename, "--save_prefix", train_prefix_2]
         
         if gpu is not None:
             args_train += ['--gpu', gpu]
-        train.command_line(arguments = args_train)
+        main(arguments = args_train)
         
         config_filename2 = train_prefix_2 + ".train.config"
