+++ tests/utils_test.py
@@ -15,6 +15,8 @@ import chainer.links as L
 import nmt_chainer.models as models
 import nmt_chainer.utilities.utils as utils
 
+
+from nmt_chainer.__main__ import main
 from nmt_chainer.utilities.utils import de_batch
 
 class TestDeBatch:
@@ -79,13 +81,13 @@ class TestMakeData:
         train_dir = tmpdir.mkdir("train")
         data_prefix = str(train_dir.join("test1.data"))
         train_prefix = str(train_dir.join("test1.train"))
-        args = [os.path.join(test_data_dir, "src2.txt"), os.path.join(test_data_dir, "tgt2.txt"), data_prefix ]
-        make_data.cmdline(arguments = args)
+        args = ["make_data", os.path.join(test_data_dir, "src2.txt"), os.path.join(test_data_dir, "tgt2.txt"), data_prefix ]
+        main(arguments = args)
         
-        args_train = [data_prefix, train_prefix] + "--max_nb_iters 5 --mb_size 2 --Ei 10 --Eo 12 --Hi 30 --Ha 70 --Ho 15 --Hl 23".split(" ")
+        args_train = ["train", data_prefix, train_prefix] + "--max_nb_iters 5 --mb_size 2 --Ei 10 --Eo 12 --Hi 30 --Ha 70 --Ho 15 --Hl 23".split(" ")
         if gpu is not None:
             args_train += ['--gpu', gpu]
-        train.command_line(arguments = args_train)
+        main(arguments = args_train)
